Fix radius input type from float to int for area calculation

+++ lab 1/task4.py
@@ -1,5 +1,5 @@
 n = input("Enter The radius of the circle : ")
-radius = float(n)
+radius = int(n)
 area = 3.14 * radius ** 2
 circumference = 2 * 3.14159 * radius
 print("Area of the circle:", area)
